requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Django==4.1.3
django-crontab==0.7.1
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
psycopg2-binary==2.9.5
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
